requirements.txt
boto3
AWSIoTPythonSDK
argparse
psutil
Pillow